% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACS_ADS_DE.R
\name{ACS_ADS_DE_plot}
\alias{ACS_ADS_DE_plot}
\title{ACS/ADS scores and DE scores visualization: Plot of pathways average DE posterior mean whose sizes are related to ACS/ADS value.}
\usage{
ACS_ADS_DE_plot(
  mcmc.merge.list,
  ACS_ADS_pathway,
  dataset.names,
  cluster.lb = NULL,
  select.pathway.list,
  highlight.pathway.index = NULL,
  plot.path = NULL
)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{ACS_ADS_pathway:}{a list of four data frames: pathway specific ACS values, ADS values
and their permuted p-value (pathway on rows, column being ACS/ADS value or the p-values).}

\item{dataset.names:}{a vector of dataset names.}

\item{cluster.lb:}{a vector of cluster label named by pathway names. optional.}

\item{select.pathway.list:}{a list of selected pathways (containing gene components).}

\item{highlight.pathway.index:}{a numeric vector indicating which pathways in the select.pathway.list need to be highlighted.}
}
\value{
A collection of ACS/ADS-DE plots in each data pair. In each individual plot, each pathway is a dot sized by ACS/ADS scores
and x-axis and y-axis are the average absolute posterior DE values. Pathways size are determined by ACS and ADS in the upper-right trianglar plots
and the lower-left triangular plots respectively.
}
\description{
ACS/ADS scores and DE scores visualization: Plot of pathways average DE posterior mean whose sizes are related to ACS/ADS value.
The \code{ACS_ADS_DE_plot} is function to visualize ACS/ADS scores and DE scores together in one plot.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
#select.pathway.list from the pathSelect step (see the example in function 'pathSelect')
#ACS_ADS_pathway from the multi_ACS_ADS_pathway step (see the example in 'multi_ACS_ADS_pathway')
#highlight and label the first three pathways in selec.pathway.list:
highlight.pathway.index = c(1,2,3)
ACS_ADS_DE_plot = function(mcmc.merge.list,ACS_ADS_pathway,dataset.names,
                           cluster.lb=NULL,select.pathway.list,
                           highlight.pathway.index=highlight.pathway.index,plot.path=NULL)
}
}
