% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes}
\alias{bayes}
\title{Run Bayesian analysis for individual pdata}
\usage{
bayes(pData, seed = 12345)
}
\arguments{
\item{pData:}{individual pData. The data matrix has to consist of
two columns, first being the p-value, second being the log fold change.}

\item{seed:}{seed number.}
}
\value{
an MCMC output matrix of signed DE indicator (input for
resemblance analysis)
}
\description{
Run Bayesian analysis for individual pData
The \code{bayes} is function to run Bayesian analysis for individual pdata
}
\examples{
\dontrun{
data(hb)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
hb_pData = summaryDE[,c(3,1)]
hb_MCMCout = bayes(hb_pData, seed=12345)
}
}
