% Generated by roxygen2 (4.0.1): do not edit by hand
\name{camo}
\alias{camo}
\title{Load a given version of a package, and return it as an object}
\usage{
camo(package, version)
}
\arguments{
\item{package}{Package name.}

\item{version}{Package version.}
}
\description{
Load a given version of a package, and return it as an object
}

