% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustDiag.R
\name{clustDiag}
\alias{clustDiag}
\title{Perform consensus clustering diagnostics to determine optimal K for pathway 
clustering.}
\usage{
clustDiag(ARS_pathway)
}
\arguments{
\item{ARS_pathway:}{a list of two data frames: pathway specific ARS values and 
their permuted p-value (pathway on rows, column being ARS value or the p-values).}
}
\value{
stored output in the folder named "clustDiag".
}
\description{
Perform consensus clustering diagnostics to determine optimal K for pathway 
clustering 
The \code{clustDiag} is function to perform consensus clustering diagnostics for 
pathway clustering. It will generate consensus CDF and delta area plot and saved to 
the folder named "clustDiag". These plots will help you determine the optimal number
of clusters K for pathway clustering. Run this first before \code{multiOutput} when 
"clustPathway" output is chosen.
}
\examples{
\dontrun{
#ARS_pathway from the multiARS step 
results <- clustDiag(ARS_pathway)
}
}
