% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indDE.R
\name{indDE}
\alias{indDE}
\title{Differential expression analysis for individual data.}
\usage{
indDE(data, group, data.type, case.label, ctrl.label)
}
\arguments{
\item{data:}{the raw expression data.}

\item{group:}{the group label.}

\item{data.type:}{either "microarray" or "RNAseq"}

\item{case.label:}{label for the case group.}

\item{control.label:}{label for the control group.}
}
\value{
summary consisting of log fold change, lfc standard error,
p-value, q-value.
}
\description{
Differential expression analysis for individual data
The \code{indDE} is function to perform differential expression
analysis for individual data
}
\examples{
\dontrun{
data(hb)
summaryDE = indDE(data=data,group=as.factor(group),data.type="microarray",
                   case.label="2", ctrl.label="1")
}
}
