% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsGlobal.R
\name{mdsGlobal}
\alias{mdsGlobal}
\title{Global MDS plot for multiple pairs: multidimensional scaling plot for global ACS or ADS scores}
\usage{
mdsGlobal(acs.value, model.name, sep = "_", file)
}
\arguments{
\item{acs.value:}{ACS or ADS scores (ACS is recommended)}

\item{model.name:}{a vector of dataset names.}

\item{sep:}{a character string to separate the data terms.}

\item{file:}{the output file name.}
}
\value{
A MDS plot of ACS or ADS scores for all DTS's to visualize global distance between each data pair.
}
\description{
Global MDS plot for multiple pairs: multidimensional scaling plot for global ACS or ADS scores
The \code{mdsGlobal} is function to generate global MDS plot for multiple pairs
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
#ACS_ADS_global from the multi_ACS_ADS_global step (see the example in 'multi_ACS_ADS_global')
dataset.names = c("hb","hs","ht","ha","hi","hl",
                  "mb","ms","mt","ma","mi","ml")
acs.value = c(ACS_ADS_global$ACS)
names(acs.value) = c(sapply(rownames(ACS_ADS_global$ACS),function(x) {
paste(x,rownames(ACS_ADS_global$ACS),sep="_")
}))
mdsGlobal(acs.value,dataset.names,sep="_",file="~/globalMDS.pdf")
}
}
