% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMCMC.R
\name{mergeMCMC}
\alias{mergeMCMC}
\title{Merge multiple MCMCout datasets preparing for cross-species
analysis}
\usage{
mergeMCMC(mcmc.list, species, ortholog.db, ortholog.file = NULL,
  reference = 1)
}
\arguments{
\item{mcmc.list:}{a list of MCMC output matrices.}

\item{species:}{a vector specie names of same length as mcmc.list.}

\item{ortholog.db:}{the ortholog object (in R environment)}

\item{ortholog.file:}{the ortholog file to be imported}

\item{reference:}{the index of the reference data, the outputted merged 
list will be named using the rownames of this data.}
}
\value{
an merged list of multiple MCMCout datasets (with same number of 
rows and rownames)
}
\description{
Merge multiple MCMCout datasets
The \code{mergeMCMC} is function to merge multiple MCMCout datasets by
matching orthologs
}
\examples{
\dontrun{
data(hb)
summaryDE <- indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
hb_pData <- summaryDE[,c(3,1)]
hb_MCMCout <- bayes(hb_pData, seed=12345)
data(hs)
summaryDE <- indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
hs_pData <- summaryDE[,c(3,1)]
hs_MCMCout <- bayes(hs_pData, seed=12345)
data(ht)
summaryDE <- indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
ht_pData <- summaryDE[,c(3,1)]
ht_MCMCout <- bayes(ht_pData, seed=12345)
data(mb)
summaryDE <- indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
mb_pData <- summaryDE[,c(3,1)]
mb_MCMCout <- bayes(mb_pData, seed=12345)
data(ms)
summaryDE <- indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
ms_pData <- summaryDE[,c(3,1)]
ms_MCMCout <- bayes(ms_pData, seed=12345)
data(mt)
summaryDE <- indDE(data=data,group=as.factor(group),data.type="microarray",
                  case.label="2", ctrl.label="1")
mt_pData <- summaryDE[,c(3,1)]
mt_MCMCout <- bayes(mt_pData, seed=12345)

#1. single pair example
mcmc.list <- list(hb_MCMCout,mb_MCMCout)
species <- c("human","mouse")
data(hm_orth)
mcmc.merge.list <- mergeMCMC(mcmc.list,species = species,
ortholog.db = hm_orth, reference=1) 

#2. multiple pairs example
mcmc.list <- list(hb_MCMCout,hs_MCMCout,ht_MCMCout,
                  mb_MCMCout,ms_MCMCout,mt_MCMCout)
species <- c(rep("human",3), rep("mouse",3))
data(hm_orth)
mcmc.merge.list <- mergeMCMC(mcmc.list,species = species,
ortholog.db = hm_orth, reference=1)                
}
}
