% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeRaw.R
\name{mergeRaw}
\alias{mergeRaw}
\title{Merge multiple raw datasets preparing for cross-species
analysis}
\usage{
mergeRaw(data.list, species, ortholog.db, ortholog.file = NULL,
  reference = 1, unique = T)
}
\arguments{
\item{data.list:}{a list of data matrices.}

\item{species:}{a vector specie names of same length as data.list.}

\item{ortholog.db:}{the ortholog object (in R environment)}

\item{ortholog.file:}{the ortholog file to be imported}

\item{reference:}{the index of the reference data, the outputted merged 
list will be named using the rownames of this data.}

\item{unique:}{logical value indicating whether to take only the "one to one" unique match (T) or allow "multiple to one" and "one to multiple" match (F)}
}
\value{
an merged list of rdatasets (with same number of 
rows and rownames)
}
\description{
Merge multiple raw datasets
The \code{mergeRaw} is function to merge multiple raw datasets by
matching orthologs
}
\examples{
\dontrun{
data(hb)
hb.data <- data
hb.group <- group
data(hs)
hs.data <- data
hs.group <- group
data(ht)
ht.data <- data
ht.group <- group
data(mb)
mb.data <- data
mb.group <- group
data(ms)
ms.data <- data
ms.group <- group
data(mt)
mt.data <- data
mt.group <- group
data.list <- list(hb.data,hs.data,ht.data,mb.data,ms.data,mt.data)
species <- c(rep("human",3), rep("mouse",3))
data(hm_orth)
data.merge.list <- mergeRaw(data.list,species=species,
ortholog.db = hm_orth, reference=1,unique=T)
}
}
