% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ACS_ADS_global.R
\name{multi_ACS_ADS_global}
\alias{multi_ACS_ADS_global}
\title{Resemblance analysis for multiple pairs: global ACS, ADS and their permuted p-value}
\usage{
multi_ACS_ADS_global(
  mcmc.merge.list,
  dataset.names,
  measure = "Fmeasure",
  B = 100
)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{dataset.names:}{a vector of dataset names.}

\item{measure:}{three types of ACS/ADS measures to be used: "youden",
"Fmeasure","geo.mean". Default is "Fmeasure".}

\item{B:}{number of permutations.}
}
\value{
Four lists: global ACS values and its permuted p-value.
global ADS values and its permuted p-value.In addition,
the four data matrices are written to a folder named "arsGlobal".
}
\description{
Resemblance analysis for multiple pairs: global ACS, ADS and their permuted p-value
The \code{multi_ACS_ADS_global} is function to perform resemblance analysis
for multiple pairs, generating global ACS, ADS and their permuted p-value.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
dataset.names = c("hb","hs","ht","ha","hi","hl",
                  "mb","ms","mt","ma","mi","ml")
ACS_ADS_global = multi_ACS_ADS_global(mcmc.merge.list,dataset.names,
measure="Fmeasure",B=1000)
}
}
