% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ACS_ADS_pathway.R
\name{multi_ACS_ADS_pathway}
\alias{multi_ACS_ADS_pathway}
\title{Resemblance analysis for multiple pairs: pathway-specific ACS, ADS and their permuted p-value}
\usage{
multi_ACS_ADS_pathway(
  mcmc.merge.list,
  dataset.names,
  select.pathway.list,
  measure = "Fmeasure",
  B = 100,
  parallel = F,
  n.cores = 4
)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{dataset.names:}{a vector of dataset names.}

\item{measure:}{three types of ACS/ADS measures to be used: "youden",
"Fmeasure","geo.mean". Default is "Fmeasure".}

\item{B:}{number of permutations.}

\item{parallel:}{whether to perform parallel ('mclapply' will be used if parallel=T) computing in permutation.}

\item{n.cores:}{if parallel=T, the number of cores to use (mc.cores parameter in 'mclapply' function).}
}
\value{
Four lists: pathway-specific ACS values and its permuted p-value.
pathway-specific ADS values and its permuted p-value.In addition,
the four data matrices are written to a folder named "arsPathway".
}
\description{
Resemblance analysis for multiple pairs: pathway-specific ACS, ADS and their
permuted p-value
The \code{multi_ACS_ADS_pathway} is function to perform resemblance analysis
for multiple pairs, generating pathway-specific ACS, ADS and their permuted p-value.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
#select.pathway.list from the pathSelect step (see the example in function 'pathSelect')
dataset.names = c("hb","hs","ht","ha","hi","hl",
                  "mb","ms","mt","ma","mi","ml")
ACS_ADS_pathway = multi_ACS_ADS_pathway(mcmc.merge.list,dataset.names,
                                        select.pathway.list,
                                        measure="Fmeasure",
                                        B=1000,parallel=F)
}
}
