% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiARS.R
\name{multiARS_global}
\alias{multiARS_global}
\title{Resemblance analysis for multiple pairs: global ARS and its 
permuted p-value.}
\usage{
multiARS_global(mcmc.merge.list, dataset.names, measure = "Fmeasure",
  parallel = F, cpu = 2, B = 50)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{dataset.names:}{a vector of dataset names.}

\item{measure:}{three types of ARS measures to be used: "youden",
"Fmeasure","geo.mean". Default is "Fmeasure".}

\item{parallel:}{whether to perform parallel computing in permutation.}

\item{cpu:}{if parallel=T, how many cpus to be used.}

\item{B:}{number of permutations.}
}
\value{
two lists: global ARS values and its permuted p-value, in addition,
the two data matrices are written to the folder named "arsGlobal".
}
\description{
Resemblance analysis for multiple pairs: global ARS and its permuted p-value
The \code{multiARS_global} is function to perform resemblance analysis 
for multiple pairs, generating global ARS and its permuted p-value.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (example 1)
dataset.names <- c("hb","hs","ht","mb","ms","mt")
ARS_global <- multiARS_global(mcmc.merge.list,dataset.names,B=100)
}
}
