% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiARS.R
\name{multiARS_pathway}
\alias{multiARS_pathway}
\title{Resemblance analysis for multiple pairs: pathway specific ARS and 
their permuted p-value.}
\usage{
multiARS_pathway(mcmc.merge.list, dataset.names, select.pathway.list,
  measure = "Fmeasure", parallel = F, cpu = 2, B = 50)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{dataset.names:}{a vector of dataset names.}

\item{select.pathway.list:}{a list of selected pathways (containing gene 
components).}

\item{measure:}{three types of ARS measures to be used: "youden",
"Fmeasure","geo.mean". Default is "Fmeasure".}

\item{parallel:}{whether to perform parallel computing in permutation.}

\item{cpu:}{if parallel=T, how many cpus to be used.}

\item{B:}{number of permutations.}
}
\value{
a list of two data frames: pathway specific ARS values and their 
permuted p-value (pathway on rows, column being ARS value or the 
p-values), in addition, both are written to the folder named 
"arsPathway".
}
\description{
Resemblance analysis for multiple pairs: pathway specific ARS and their 
permuted p-value
The \code{multiARS_pathway} is function to perform resemblance analysis 
for multiple pairs, generating pathway specific ARS and their permuted 
p-value.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (example 2)
#select.pathway.list from the pathSelect step
dataset.names <- c("hb","hs","ht","mb","ms","mt")
ARS_pathway <- multiARS_pathway(mcmc.merge.list,dataset.names,
select.pathway.list,B=100)
}
}
