% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOutput.R
\name{multiOutput}
\alias{multiOutput}
\title{Downstream visualization tools}
\usage{
multiOutput(
  mcmc.merge.list,
  dataset.names,
  select.pathway.list,
  ACS_ADS_pathway,
  output = c("clustPathway", "mdsModel", "clustModel", "genePM", "keggView",
    "reactomeView", "comemberPlot"),
  optK = NULL,
  sil_cut = 0.1,
  ADS_cluster = FALSE,
  hashtb = NULL,
  pathways = NULL,
  keywords_cut = 0.05,
  text.permutation = "all",
  comemberProb_cut = 0.7,
  ViewPairSelect = NULL,
  kegg.species = "hsa",
  KEGG.genes.entrezID = NULL,
  KEGG.pathID2name = NULL,
  reactome.species = "HSA",
  Reactome.genes.topologyGtype = NULL,
  Reactome.pathID2name = NULL
)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{dataset.names:}{a vector of dataset names.}

\item{select.pathway.list:}{a list of selected pathways (containing gene components).}

\item{ACS_ADS_pathway:}{a list of four data frames: pathway specific ACS values, ADS values
and their permuted p-value (pathway on rows, column being ACS/ADS value or the p-values).}

\item{output:}{seven options: "clustPathway" (pathway clustering),"comemberPlot" (heatmaps of
probabilities that each model pair been clustered together in pathways in each pathway cluster),
"mdsModel"(model MDS plot),"clustModel" (model clustering output), "genePM" (heatmap of gene
posterior mean), "keggView" (KEGG pathway topology, default is human - hsa, for other species,
KEGG organism name and gene Entrez ID needs to be provided as 'KEGG.genes.entrezID'),"reactomeView"
(Reactome pathway topology, default is human - HSA, for other species, Reactome organism name needs
to be provided as "reactome.species"). In single pair case (2 DTS's), clustering analysis is not
applicable and only 'genePM','keggView' can be generated. For details, please refer to manuscript.
"output" cannot be empty.}

\item{optK:}{Optimal number of clusters based on pathway clustering diagnostic results.For
"clustPathway" output only.}

\item{sil_cut:}{silhouette index to control scatterness. Larger value indicates tigher cluster and
more scattered pathways.}

\item{ADS_cluster:}{whether clustered by ADS in clustPathway. Default is FALSE.}

\item{hashtb:}{a flat hash table for text mining. Two prepared hastb table are provided: "hashtb_human"
and "hashtb_worm". Please refer to Zeng, Xiangrui, et al. "Comparative Pathway Integrator: a framework
of meta-analytic integration of multiple transcriptomic studies for consensual and differential pathway
analysis." Genes 11.6 (2020): 696.}

\item{pathways:}{complete pathway names for text mining.}

\item{keywords_cut:}{keywords above this cut will be shown in text mining spreadsheet output.}

\item{text.permutation:}{select from "all" or "enriched". In text mining, "all" permutates pathways from
full pathway.list provided while "enriched" permutates from selected pathways. "all" is suitable for
cross-species comparision while "enriched" is recommended for within-species comparision.}

\item{ViewPairSelect:}{which two datasets to view in KEGG/Reactome topology. All pairs will be
considered under default (may take a while).}

\item{comemberProb_cut:}{probability below this cut will be colored blue in comembership heatmaps.}

\item{kegg.species:}{KEGG species abbreviation. For "keggView" only. Default is "hsa".}

\item{KEGG.genes.entrezID:}{a data frame which maps gene names in mcmc.merge.list (first column) to
Entrez IDs (second column). Only required for non-human genes after merging. For "keggView" only.}

\item{KEGG.pathID2name:}{a list where each element is a KEGG pathway name with correponding pathway ID
as its name. ID will be retrieved from KEGGREST if this is NULL.}

\item{reactome.species:}{Reactome species abbreviation. For "reactomeView" only. Default is "HSA".}

\item{Reactome.genes.topologyGtype:}{a data frame which maps gene names in mcmc.merge.list (first
column) to gene name types in Reactome topology (second column). For "reactomeView" only.}

\item{Reactome.pathID2name:}{a list where each element is a Reactome pathway name with correponding
pathway ID as its name. ID will be retrieved from reactome.db if this is NULL.}
}
\value{
stored output in created folders.
}
\description{
Downstream visualization tools: visualization outputs including overall
pathway clustering and output for each pathway
The \code{multiOutput} is function to visualize pathway ACS/ADS results: including overall
pathway clustering outputs, comembership heatmaps, model MDS plot, model clustering output,
heatmap of gene posterior mean, kegg pathway topology for each pathway.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step (see the example in function 'merge')
#select.pathway.list from the pathSelect step (see the example in function 'pathSelect')
#ACS_ADS_pathway from the multi_ACS_ADS_pathway step (see example in 'multi_ACS_ADS_pathway')
data(hashtb_human) #include hashtb & pathways for text mining
dataset.names = c("hb","hs","ht","ha","hi","hl",
                  "mb","ms","mt","ma","mi","ml")
#1. step1: select K by elbow plot from consensus clustering
ACSpvalue.mat = ACS_ADS_pathway$ACSpvalue.mat
results = clustPathway(ACSpvalue.mat)

#2. step2: run multiOutput with pre-selected K
multiOutput(mcmc.merge.list,dataset.names,select.pathway.list,ACS_ADS_pathway,
           output=c("clustPathway","comemberList","mdsModel","clustModel","genePM","keggView"),
           hashtb=hashtb,pathways=pathways,optK = K,thres = 0.2)
}
}
