% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleARS.R
\name{singleARS_global}
\alias{singleARS_global}
\title{Resemblance analysis for single pair: global ARS and its permuted 
p-value.}
\usage{
singleARS_global(mcmc.merge.list, measure = "Fmeasure", parallel = F,
  cpu = 2, B = 50)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{measure:}{three types of ARS measures to be used: "youden",
"Fmeasure","geo.mean". Default is "Fmeasure".}

\item{parallel:}{whether to perform parallel computing in permutation.}

\item{cpu:}{if parallel=T, how many cpus to be used.}

\item{B:}{number of permutations.}
}
\value{
global ARS values and its permuted p-value, in addition, the two 
values are written to the folder named "arsGlobal".
}
\description{
Resemblance analysis for single pair: global ARS and its permuted p-value
The \code{singleARS_global} is function to perform resemblance analysis 
for single pair, generating global ARS and its permuted p-value.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step
ARS_global <- singleARS_global(mcmc.merge.list,B=100)
}
}
