% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleARS.R
\name{singleARS_pathway}
\alias{singleARS_pathway}
\title{Resemblance analysis for single pair: pathway specific ARS and 
their permuted p-value.}
\usage{
singleARS_pathway(mcmc.merge.list, select.pathway.list, measure = "Fmeasure",
  parallel = F, cpu = 2, B = 50)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{select.pathway.list:}{a list of selected pathways (containing gene 
components).}

\item{measure:}{three types of ARS measures to be used: "youden",
"Fmeasure","geo.mean". Default is "Fmeasure".}

\item{parallel:}{whether to perform parallel computing in permutation.}

\item{cpu:}{if parallel=T, how many cpus to be used.}

\item{B:}{number of permutations.}
}
\value{
a data frame of pathway specific ARS values and their permuted 
p-value (pathway on rows, 1st column being ARS value and 2nd column being
the p-values), in addition, the dataframe is written to the folder named 
"arsPathway".
}
\description{
Resemblance analysis for single pair: pathway specific ARS and their 
permuted p-value
The \code{singleARS_pathway} is function to perform resemblance analysis 
for single pair, generating pathway specific ARS and their permuted 
p-value.
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step
#select.pathway.list from the pathSelect step
ARS_pathway <- singleARS_pathway(mcmc.merge.list,select.pathway.list,B=100)
}
}
