% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleOutput.R
\name{singleOutput}
\alias{singleOutput}
\title{Analysis results for single pair: visualization outputs}
\usage{
singleOutput(mcmc.merge.list, dataset.names, select.pathway.list,
  output = c("genePM", "keggView"), kegg_pathname = NULL,
  hs_gene_id = NULL)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{select.pathway.list:}{a list of selected pathways (containing gene 
components).}

\item{dataset.names:}{a vector of dataset names.}

\item{output:}{two options: "genePM" (generating heatmap of gene 
posterior mean),"keggView" (generating kegg pathway topology, human KEGG only). 
cannot be empty.}

\item{kegg_pathname:}{KEGG pathway name list. For "keggView" only.}

\item{hs_gene_id:}{Human sapiens gene id. For "keggView" only.}
}
\value{
stored output in created folders.
}
\description{
Analysis results for single pair: visualization outputs for each pathway
The \code{singleOutput} is function to generate visualization outputs 
for single pair: including heatmap of gene posterior mean, kegg 
pathway topology for each pathway
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step
#select.pathway.list from the pathSelect step
dataset.names = c("hb","mb")
library(KEGG.db)
kegg_pathname <- unlist(as.list(KEGGPATHID2NAME))
library("org.Hs.eg.db")
hs_gene_id <- unlist(mget(x=rownames(mcmc.merge.list[[1]]),
envir=org.Hs.egALIAS2EG))
singleOutput(mcmc.merge.list,dataset.names,select.pathway.list,
output=c("genePM","keggView"))
}
}
