% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOutput.R
\name{multiOutput}
\alias{multiOutput}
\title{Analysis results for multiple pairs: visualization outputs}
\usage{
multiOutput(mcmc.merge.list, dataset.names, select.pathway.list, ARS_pathway,
  output = c("clustPathway", "mdsModel", "clustModel", "genePM", "keggView"),
  hashtb = NULL, pathways = NULL, keggViewSelect = c(1, 2), optK = NULL,
  kegg_pathname = NULL, hs_gene_id = NULL)
}
\arguments{
\item{mcmc.merge.list:}{a list of merged MCMC output matrices.}

\item{dataset.names:}{a vector of dataset names.}

\item{select.pathway.list:}{a list of selected pathways (containing gene 
components).}

\item{ARS_pathway:}{a list of two data frames: pathway specific ARS values and 
their permuted p-value (pathway on rows, column being ARS value or the p-values).}

\item{output:}{five options: "clustPathway" (pathway clustering),"mdsModel"(model 
MDS plot),"clustModel" (model clustering output), "genePM" (generating heatmap of 
gene posterior mean),"keggView" (generating kegg pathway topology, human KEGG 
only). For details, please refer to manuscript. cannot be empty.}

\item{hashtb:}{hash table for text mining.}

\item{pathways:}{complete pathway names for text mining.}

\item{keggViewSelect:}{which two datasets to view in KEGG topology.}

\item{optK:}{Optimal number of clusters based on clustering diagnostic results. For
"clustPathway" output only.}

\item{kegg_pathname:}{KEGG pathway name list. For "keggView" only.}

\item{hs_gene_id:}{Human sapiens gene id. For "keggView" only.}
}
\value{
stored output in created folders.
}
\description{
Analysis results for multiple pairs: visualization outputs including overall 
pathway clustering and output for each pathway
The \code{multiOutput} is function to generate visualization outputs 
for multiple pairs: including overall pathway clustering outputs, model MDS plot, 
model clustering output, heatmap of gene posterior mean, kegg pathway topology for 
each pathway
}
\examples{
\dontrun{
#mcmc.merge.list from the merge step
#select.pathway.list from the pathSelect step
#ARS_pathway from the multiARS step
data(hashtb) #include hashtb & pathways
dataset.names <- c("hb","hs","ht","mb","ms","mt")
library(KEGG.db)
kegg_pathname <- unlist(as.list(KEGGPATHID2NAME))
library("org.Hs.eg.db")
hs_gene_id <- unlist(mget(x=rownames(mcmc.merge.list[[1]]),
envir=org.Hs.egALIAS2EG))
multiOutput(mcmc.merge.list,dataset.names,select.pathway.list,
ARS_pathway, output=c("clustPathway","mdsModel","clustModel","genePM","keggView"),
hashtb=hashtb,pathways=pathways,keggViewSelect = c(1,4),optK=7)
}
}
